@@ -1,17 +1,15 @@
 package middleware
 
 import (
-	"log"
-	"reflect"
-
-	"github.com/keecon/protoactor-go/actor"
+	"github.com/asynkron/protoactor-go/actor"
+	"log/slog"
 )
 
 // Logger is message middleware which logs messages before continuing to the next middleware.
 func Logger(next actor.ReceiverFunc) actor.ReceiverFunc {
 	fn := func(c actor.ReceiverContext, env *actor.MessageEnvelope) {
 		message := env.Message
-		log.Printf("%v got %v %+v", c.Self(), reflect.TypeOf(message), message)
+		c.Logger().Info("Actor got message", slog.Any("self", c.Self()), slog.Any("message", message))
 		next(c, env)
 	}
 
