@@ -1,14 +1,16 @@
 package cluster
 
 import (
+	"log/slog"
 	"time"
+
+	"google.golang.org/protobuf/types/known/emptypb"
 
 	"github.com/asynkron/gofun/set"
 
-	"github.com/keecon/protoactor-go/actor"
-	"github.com/keecon/protoactor-go/extensions"
-	"github.com/keecon/protoactor-go/log"
-	"github.com/keecon/protoactor-go/remote"
+	"github.com/asynkron/protoactor-go/actor"
+	"github.com/asynkron/protoactor-go/extensions"
+	"github.com/asynkron/protoactor-go/remote"
 )
 
 var extensionID = extensions.NextExtensionID()
@@ -16,7 +18,8 @@ var extensionID = extensions.NextExtensionID()
 type Cluster struct {
 	ActorSystem    *actor.ActorSystem
 	Config         *Config
-	Gossip         Gossiper
+	Gossip         *Gossiper
+	PubSub         *PubSub
 	Remote         *remote.Remote
 	PidCache       *PidCacheValue
 	MemberList     *MemberList
@@ -44,6 +47,7 @@ func New(actorSystem *actor.ActorSystem, config *Config) *Cluster {
 
 	var err error
 	c.Gossip, err = newGossiper(c)
+	c.PubSub = NewPubSub(c)
 
 	if err != nil {
 		panic(err)
@@ -87,7 +91,7 @@ func (c *Cluster) StartMember() {
 	c.Remote.Start()
 
 	address := c.ActorSystem.Address()
-	plog.Info("Starting Proto.Actor cluster member", log.String("id", c.ActorSystem.ID), log.String("address", address))
+	c.Logger().Info("Starting Proto.Actor cluster member", slog.String("address", address))
 
 	c.IdentityLookup = cfg.IdentityLookup
 	c.IdentityLookup.Setup(c, c.GetClusterKinds(), false)
@@ -97,6 +101,7 @@ func (c *Cluster) StartMember() {
 	if err := c.Gossip.StartGossiping(); err != nil {
 		panic(err)
 	}
+	c.PubSub.Start()
 	c.MemberList.InitializeTopologyConsensus()
 
 	if err := cfg.ClusterProvider.StartMember(c); err != nil {
@@ -122,7 +127,7 @@ func (c *Cluster) StartClient() {
 	c.Remote.Start()
 
 	address := c.ActorSystem.Address()
-	plog.Info("Starting Proto.Actor cluster-client", log.String("address", address))
+	c.Logger().Info("Starting Proto.Actor cluster-client", slog.String("address", address))
 
 	c.IdentityLookup = cfg.IdentityLookup
 	c.IdentityLookup.Setup(c, c.GetClusterKinds(), true)
@@ -130,11 +135,12 @@ func (c *Cluster) StartClient() {
 	if err := cfg.ClusterProvider.StartClient(c); err != nil {
 		panic(err)
 	}
+	c.PubSub.Start()
 }
 
 func (c *Cluster) Shutdown(graceful bool) {
+	c.Gossip.SetState(GracefullyLeftKey, &emptypb.Empty{})
 	c.ActorSystem.Shutdown()
-
 	if graceful {
 		_ = c.Config.ClusterProvider.Shutdown(graceful)
 		c.IdentityLookup.Shutdown()
@@ -148,21 +154,21 @@ func (c *Cluster) Shutdown(graceful bool) {
 	c.Remote.Shutdown(graceful)
 
 	address := c.ActorSystem.Address()
-	plog.Info("Stopped Proto.Actor cluster", log.String("address", address))
+	c.Logger().Info("Stopped Proto.Actor cluster", slog.String("address", address))
 }
 
 func (c *Cluster) Get(identity string, kind string) *actor.PID {
 	return c.IdentityLookup.Get(NewClusterIdentity(identity, kind))
 }
 
-func (c *Cluster) Request(identity string, kind string, message interface{}) (interface{}, error) {
+func (c *Cluster) Request(identity string, kind string, message interface{}, option ...GrainCallOption) (interface{}, error) {
 	return c.context.Request(identity, kind, message)
 }
 
 func (c *Cluster) GetClusterKind(kind string) *ActivatedKind {
 	k, ok := c.kinds[kind]
 	if !ok {
-		plog.Error("Invalid kind", log.String("kind", kind))
+		c.Logger().Error("Invalid kind", slog.String("kind", kind))
 
 		return nil
 	}
@@ -180,57 +186,28 @@ func (c *Cluster) initKinds() {
 	for name, kind := range c.Config.Kinds {
 		c.kinds[name] = kind.Build(c)
 	}
+	c.ensureTopicKindRegistered()
 }
 
-// Call is a wrap of context.RequestFuture with retries.
-func (c *Cluster) Call(name string, kind string, msg interface{}, opts ...GrainCallOption) (interface{}, error) {
-	callConfig := DefaultGrainCallConfig(c)
-	for _, o := range opts {
-		o(callConfig)
-	}
-
-	_context := callConfig.Context
-	if _context == nil {
-		_context = c.ActorSystem.Root
-	}
-
-	var lastError error
-
-	for i := 0; i < callConfig.RetryCount; i++ {
-		pid := c.Get(name, kind)
-
-		if pid == nil {
-			return nil, remote.ErrUnknownError
+// ensureTopicKindRegistered ensures that the topic kind is registered in the cluster
+// if topic kind is not registered, it will be registered automatically
+func (c *Cluster) ensureTopicKindRegistered() {
+	hasTopicKind := false
+	for name := range c.kinds {
+		if name == TopicActorKind {
+			hasTopicKind = true
+			break
 		}
-
-		timeout := callConfig.Timeout
-		_resp, err := _context.RequestFuture(pid, msg, timeout).Result()
-		if err != nil {
-			plog.Error("cluster.RequestFuture failed", log.Error(err), log.PID("pid", pid))
-			lastError = err
-
-			switch err {
-			case actor.ErrTimeout, remote.ErrTimeout:
-				callConfig.RetryAction(i)
-
-				id := ClusterIdentity{Kind: kind, Identity: name}
-				c.PidCache.Remove(id.Identity, id.Kind)
-
-				continue
-			case actor.ErrDeadLetter, remote.ErrDeadLetter:
-				callConfig.RetryAction(i)
-
-				id := ClusterIdentity{Kind: kind, Identity: name}
-				c.PidCache.Remove(id.Identity, id.Kind)
-
-				continue
-			default:
-				return nil, err
-			}
-		}
-
-		return _resp, nil
 	}
+	if !hasTopicKind {
+		store := &EmptyKeyValueStore[*Subscribers]{}
 
-	return nil, lastError
+		c.kinds[TopicActorKind] = NewKind(TopicActorKind, actor.PropsFromProducer(func() actor.Actor {
+			return NewTopicActor(store, c.Logger())
+		})).Build(c)
+	}
+}
+
+func (c *Cluster) Logger() *slog.Logger {
+	return c.ActorSystem.Logger()
 }
