@@ -1,6 +1,8 @@
 package actor
 
-import "github.com/keecon/protoactor-go/log"
+import (
+	"log/slog"
+)
 
 type Behavior []ReceiveFunc
 
@@ -26,7 +28,7 @@ func (b *Behavior) Receive(context Context) {
 	if ok {
 		behavior(context)
 	} else {
-		plog.Error("empty behavior called", log.Stringer("pid", context.Self()))
+		context.Logger().Error("empty behavior called", slog.Any("pid", context.Self()))
 	}
 }
 
