@@ -10,8 +10,7 @@ The following quote from Wikipedia distills the definition of an actor down to i
 	In response to a message that it receives, an actor can: make local decisions, create more actors,
 	send more messages, and determine how to respond to the next message received.
 
-
-Creating Actors
+# Creating Actors
 
 Props provide the building blocks for declaring how actors should be created. The following example defines an actor
 using a function literal to process messages:
@@ -40,8 +39,7 @@ The result of calling Spawn is a unique PID or process identifier.
 Each time an actor is spawned, a new mailbox is created and associated with the PID. Messages are sent to the mailbox
 and then forwarded to the actor to process.
 
-
-Processing Messages
+# Processing Messages
 
 An actor processes messages via its Receive handler. The signature of this function is:
 
@@ -50,7 +48,7 @@ An actor processes messages via its Receive handler. The signature of this funct
 The actor system guarantees that this method is called synchronously, therefore there is no requirement to protect
 shared state inside calls to this function.
 
-Communicating With Actors
+# Communicating With Actors
 
 A PID is the primary interface for sending messages to actors. Context.Send is used to send an asynchronous
 message to the actor associated with the PID:
@@ -67,5 +65,6 @@ For synchronous communication, an actor will use a Future and wait for the resul
 to an actor and wait for a response, use the RequestFuture method, which returns a Future:
 
 	f := actor.RequestFuture(pid,"Hello", 50 * time.Millisecond)
-	res, err := f.Result() // waits for pid to reply */
+	res, err := f.Result() // waits for pid to reply
+*/
 package actor
