@@ -1,6 +1,7 @@
 package actor
 
 import (
+	"log/slog"
 	"time"
 )
 
@@ -38,6 +39,10 @@ func (rc RootContext) Copy() *RootContext {
 
 func (rc *RootContext) ActorSystem() *ActorSystem {
 	return rc.actorSystem
+}
+
+func (rc *RootContext) Logger() *slog.Logger {
+	return rc.actorSystem.Logger()
 }
 
 func (rc *RootContext) WithHeaders(headers map[string]string) *RootContext {
@@ -170,7 +175,7 @@ func (rc *RootContext) SpawnPrefix(props *Props, prefix string) *PID {
 
 // SpawnNamed starts a new actor based on props and named using the specified name
 //
-// ErrNameExists will be returned if id already exists
+// # ErrNameExists will be returned if id already exists
 //
 // Please do not use name sharing same pattern with system actors, for example "YourPrefix$1", "Remote$1", "future$1".
 func (rc *RootContext) SpawnNamed(props *Props, name string) (*PID, error) {
