@@ -1,7 +1,7 @@
 package actor
 
 import (
-	"github.com/keecon/protoactor-go/log"
+	"log/slog"
 )
 
 // SupervisorEvent is sent on the EventStream when a supervisor have applied a directive to a failing child actor
@@ -14,7 +14,7 @@ type SupervisorEvent struct {
 func SubscribeSupervision(actorSystem *ActorSystem) {
 	_ = actorSystem.EventStream.Subscribe(func(evt interface{}) {
 		if supervisorEvent, ok := evt.(*SupervisorEvent); ok {
-			plog.Debug("[SUPERVISION]", log.Stringer("actor", supervisorEvent.Child), log.Stringer("directive", supervisorEvent.Directive), log.Object("reason", supervisorEvent.Reason))
+			actorSystem.Logger().Debug("[SUPERVISION]", slog.Any("actor", supervisorEvent.Child), slog.Any("directive", supervisorEvent.Directive), slog.Any("reason", supervisorEvent.Reason))
 		}
 	})
 }
