@@ -5,11 +5,11 @@ package cluster
 import (
 	"context"
 	"fmt"
+	"log/slog"
 	"time"
 
-	"github.com/keecon/protoactor-go/actor"
-	"github.com/keecon/protoactor-go/log"
-	"github.com/keecon/protoactor-go/remote"
+	"github.com/asynkron/protoactor-go/actor"
+	"github.com/asynkron/protoactor-go/remote"
 )
 
 // Defines a type to provide DefaultContext configurations / implementations.
@@ -32,30 +32,32 @@ func newDefaultClusterContext(cluster *Cluster) Context {
 	return &clusterContext
 }
 
-func (dcc *DefaultContext) Request(identity, kind string, message interface{}, timeout ...time.Duration) (interface{}, error) {
+func (dcc *DefaultContext) Request(identity, kind string, message interface{}, opts ...GrainCallOption) (interface{}, error) {
 	var err error
 
 	var resp interface{}
 
 	var counter int
+	callConfig := DefaultGrainCallConfig(dcc.cluster)
+	for _, o := range opts {
+		o(callConfig)
+	}
+
+	_context := callConfig.Context
 
 	// get the configuration from the composed Cluster value
-	cfg := dcc.cluster.Config.ToClusterContextConfig()
+	cfg := dcc.cluster.Config.ToClusterContextConfig(dcc.cluster.Logger())
 
 	start := time.Now()
 
-	plog.Debug(fmt.Sprintf("Requesting %s:%s Message %#v", identity, kind, message))
+	dcc.cluster.Logger().Debug(fmt.Sprintf("Requesting %s:%s Message %#v", identity, kind, message))
 
 	// crate a new Timeout Context
-	ttl := cfg.ActorRequestTimeout
-	if len(timeout) > 0 {
-		ttl = timeout[0]
-	}
+	ttl := callConfig.Timeout
 
 	ctx, cancel := context.WithTimeout(context.Background(), ttl)
 	defer cancel()
 
-	_context := dcc.cluster.ActorSystem.Root
 selectloop:
 	for {
 		select {
@@ -65,23 +67,24 @@ selectloop:
 
 			break selectloop
 		default:
-			pid := dcc.getCachedPid(identity, kind)
+			pid := dcc.getPid(identity, kind)
 			if pid == nil {
-				plog.Debug(fmt.Sprintf("Requesting %s:%s did not get PID from IdentityLookup", identity, kind))
-				counter = cfg.RetryAction(counter)
-
+				dcc.cluster.Logger().Debug("Requesting PID from IdentityLookup but got nil", slog.String("identity", identity), slog.String("kind", kind))
+				counter = callConfig.RetryAction(counter)
 				continue
 			}
 
+			//TODO: why is err != nil when res != nil?
 			resp, err = _context.RequestFuture(pid, message, ttl).Result()
+			if resp != nil {
+				break selectloop
+			}
 			if err != nil {
-				plog.Error("cluster.RequestFuture failed", log.Error(err), log.PID("pid", pid))
+				dcc.cluster.Logger().Error("cluster.RequestFuture failed", slog.Any("error", err), slog.Any("pid", pid))
 				switch err {
 				case actor.ErrTimeout, remote.ErrTimeout, actor.ErrDeadLetter, remote.ErrDeadLetter:
-					counter = cfg.RetryAction(counter)
+					counter = callConfig.RetryAction(counter)
 					dcc.cluster.PidCache.Remove(identity, kind)
-					err = nil // reset our error variable as we can succeed still
-
 					continue
 				default:
 					break selectloop
@@ -97,7 +100,7 @@ selectloop:
 
 	if contextError := ctx.Err(); contextError != nil && cfg.requestLogThrottle() == actor.Open {
 		// context timeout exceeded, report and return
-		plog.Warn(fmt.Sprintf("Request retried but failed for %s:%s, elapsed %v", identity, kind, totalTime))
+		dcc.cluster.Logger().Warn("Request retried but failed", slog.String("identity", identity), slog.String("kind", kind), slog.Duration("duration", totalTime))
 	}
 
 	return resp, err
@@ -105,16 +108,12 @@ selectloop:
 
 // gets the cached PID for the given identity
 // it can return nil if none is found.
-func (dcc *DefaultContext) getCachedPid(identity, kind string) *actor.PID {
+func (dcc *DefaultContext) getPid(identity, kind string) *actor.PID {
 	pid, _ := dcc.cluster.PidCache.Get(identity, kind)
+	if pid == nil {
+		pid = dcc.cluster.Get(identity, kind)
+		dcc.cluster.PidCache.Set(identity, kind, pid)
+	}
 
 	return pid
-}
-
-// default retry action, it just sleeps incrementally.
-func defaultRetryAction(i int) int {
-	i++
-	time.Sleep(time.Duration(i * i * 50))
-
-	return i
 }
