@@ -4,7 +4,6 @@ package cluster
 
 import (
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/anypb"
 )
 
 // customary type that defines a states sender callback.
@@ -13,7 +12,7 @@ type LocalStateSender func(memberStateDelta *MemberStateDelta, member *Member)
 // This interface must be implemented by any value that.
 // wants to be used as a gossip state storage
 type GossipStateStorer interface {
-	GetState(key string) map[string]*anypb.Any
+	GetState(key string) map[string]*GossipKeyValue
 	SetState(key string, value proto.Message)
 }
 
