@@ -5,26 +5,29 @@ package cluster
 import (
 	"errors"
 	"fmt"
+	"log/slog"
+	"strings"
 	"time"
+
+	"github.com/asynkron/protoactor-go/remote"
 
 	"github.com/asynkron/gofun/set"
 	"google.golang.org/protobuf/proto"
 
-	"github.com/keecon/protoactor-go/actor"
-	"github.com/keecon/protoactor-go/log"
+	"github.com/asynkron/protoactor-go/actor"
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
 const DefaultGossipActorName string = "gossip"
 
-// Used to update gossip data when a Clustertopology event occurs
+// GossipUpdate Used to update gossip data when a ClusterTopology event occurs
 type GossipUpdate struct {
 	MemberID, Key string
 	Value         *anypb.Any
 	SeqNumber     int64
 }
 
-// Customary type used to provide consensus check callbacks of any type
+// ConsensusChecker Customary type used to provide consensus check callbacks of any type
 // note: this is equivalent to (for future go v1.18):
 //
 //	type ConsensusChecker[T] func(GossipState, map[string]empty) (bool, T)
@@ -49,9 +52,9 @@ type Gossiper struct {
 }
 
 // Creates a new Gossiper value and return it back
-func newGossiper(cl *Cluster, opts ...Option) (Gossiper, error) {
+func newGossiper(cl *Cluster, opts ...Option) (*Gossiper, error) {
 	// create a new Gossiper value
-	gossiper := Gossiper{
+	gossiper := &Gossiper{
 		GossipActorName: DefaultGossipActorName,
 		cluster:         cl,
 		close:           make(chan struct{}),
@@ -59,14 +62,14 @@ func newGossiper(cl *Cluster, opts ...Option) (Gossiper, error) {
 
 	// apply any given options
 	for _, opt := range opts {
-		opt(&gossiper)
+		opt(gossiper)
 	}
 
 	return gossiper, nil
 }
 
-func (g *Gossiper) GetState(key string) (map[string]*anypb.Any, error) {
-	plog.Debug(fmt.Sprintf("Gossiper getting state from %s", g.pid))
+func (g *Gossiper) GetState(key string) (map[string]*GossipKeyValue, error) {
+	g.cluster.Logger().Debug(fmt.Sprintf("Gossiper getting state from %s", g.pid))
 
 	msg := NewGetGossipStateRequest(key)
 	timeout := g.cluster.Config.TimeoutTime
@@ -74,13 +77,13 @@ func (g *Gossiper) GetState(key string) (map[string]*anypb.Any, error) {
 	if err != nil {
 		switch err {
 		case actor.ErrTimeout:
-			plog.Error("Could not get a response from GossipActor: request timeout", log.Error(err), log.String("remote", g.pid.String()))
+			g.cluster.Logger().Error("Could not get a response from GossipActor: request timeout", slog.Any("error", err), slog.String("remote", g.pid.String()))
 			return nil, err
 		case actor.ErrDeadLetter:
-			plog.Error("remote no longer exists", log.Error(err), log.String("remote", g.pid.String()))
+			g.cluster.Logger().Error("remote no longer exists", slog.Any("error", err), slog.String("remote", g.pid.String()))
 			return nil, err
 		default:
-			plog.Error("Could not get a response from GossipActor", log.Error(err), log.String("remote", g.pid.String()))
+			g.cluster.Logger().Error("Could not get a response from GossipActor", slog.Any("error", err), slog.String("remote", g.pid.String()))
 			return nil, err
 		}
 	}
@@ -89,17 +92,17 @@ func (g *Gossiper) GetState(key string) (map[string]*anypb.Any, error) {
 	response, ok := r.(*GetGossipStateResponse)
 	if !ok {
 		err := fmt.Errorf("could not promote %T interface to GetGossipStateResponse", r)
-		plog.Error("Could not get a response from GossipActor", log.Error(err), log.String("remote", g.pid.String()))
+		g.cluster.Logger().Error("Could not get a response from GossipActor", slog.Any("error", err), slog.String("remote", g.pid.String()))
 		return nil, err
 	}
 
 	return response.State, nil
 }
 
-// Sends fire and forget message to update member state
+// SetState Sends fire and forget message to update member state
 func (g *Gossiper) SetState(key string, value proto.Message) {
 	if g.throttler() == actor.Open {
-		plog.Debug(fmt.Sprintf("Gossiper setting state %s to %s", key, g.pid))
+		g.cluster.Logger().Debug(fmt.Sprintf("Gossiper setting state %s to %s", key, g.pid))
 	}
 
 	if g.pid == nil {
@@ -110,10 +113,10 @@ func (g *Gossiper) SetState(key string, value proto.Message) {
 	g.cluster.ActorSystem.Root.Send(g.pid, &msg)
 }
 
-// Sends a Request (that blocks) to update member state
+// SetStateRequest Sends a Request (that blocks) to update member state
 func (g *Gossiper) SetStateRequest(key string, value proto.Message) error {
 	if g.throttler() == actor.Open {
-		plog.Debug(fmt.Sprintf("Gossiper setting state %s to %s", key, g.pid))
+		g.cluster.Logger().Debug(fmt.Sprintf("Gossiper setting state %s to %s", key, g.pid))
 	}
 
 	if g.pid == nil {
@@ -124,10 +127,10 @@ func (g *Gossiper) SetStateRequest(key string, value proto.Message) error {
 	r, err := g.cluster.ActorSystem.Root.RequestFuture(g.pid, &msg, g.cluster.Config.TimeoutTime).Result()
 	if err != nil {
 		if err == actor.ErrTimeout {
-			plog.Error("Could not get a response from Gossiper Actor: request timeout", log.String("remote", g.pid.String()))
+			g.cluster.Logger().Error("Could not get a response from Gossiper Actor: request timeout", slog.String("remote", g.pid.String()))
 			return err
 		}
-		plog.Error("Could not get a response from Gossiper Actor", log.Error(err), log.String("remote", g.pid.String()))
+		g.cluster.Logger().Error("Could not get a response from Gossiper Actor", slog.Any("error", err), slog.String("remote", g.pid.String()))
 		return err
 	}
 
@@ -135,7 +138,7 @@ func (g *Gossiper) SetStateRequest(key string, value proto.Message) error {
 	_, ok := r.(*SetGossipStateResponse)
 	if !ok {
 		err := fmt.Errorf("could not promote %T interface to SetGossipStateResponse", r)
-		plog.Error("Could not get a response from Gossip Actor", log.Error(err), log.String("remote", g.pid.String()))
+		g.cluster.Logger().Error("Could not get a response from Gossip Actor", slog.Any("error", err), slog.String("remote", g.pid.String()))
 		return err
 	}
 	return nil
@@ -148,19 +151,19 @@ func (g *Gossiper) SendState() {
 
 	r, err := g.cluster.ActorSystem.Root.RequestFuture(g.pid, &SendGossipStateRequest{}, 5*time.Second).Result()
 	if err != nil {
-		plog.Warn("Gossip could not send gossip request", log.PID("PID", g.pid), log.Error(err))
+		g.cluster.Logger().Warn("Gossip could not send gossip request", slog.Any("PID", g.pid), slog.Any("error", err))
 		return
 	}
 
 	if _, ok := r.(*SendGossipStateResponse); !ok {
-		plog.Error("Gossip SendState received unknown response", log.Message(r))
+		g.cluster.Logger().Error("Gossip SendState received unknown response", slog.Any("message", r))
 	}
 }
 
-// Builds a consensus handler and a consensus checker, send the checker to the
+// RegisterConsensusCheck Builds a consensus handler and a consensus checker, send the checker to the
 // Gossip actor and returns the handler back to the caller
 func (g *Gossiper) RegisterConsensusCheck(key string, getValue func(*anypb.Any) interface{}) ConsensusHandler {
-	definition := NewConsensusCheckBuilder(key, getValue)
+	definition := NewConsensusCheckBuilder(g.cluster.Logger(), key, getValue)
 	consensusHandle, check := definition.Build()
 	request := NewAddConsensusCheck(consensusHandle.GetID(), check)
 	g.cluster.ActorSystem.Root.Send(g.pid, &request)
@@ -169,7 +172,7 @@ func (g *Gossiper) RegisterConsensusCheck(key string, getValue func(*anypb.Any) 
 
 func (g *Gossiper) StartGossiping() error {
 	var err error
-	g.pid, err = g.cluster.ActorSystem.Root.SpawnNamed(actor.PropsFromProducer(func() actor.Actor {
+	g.pid, err = g.cluster.ActorSystem.Root.SpawnNamed(actor.PropsFromProducerWithActorSystem(func(system *actor.ActorSystem) actor.Actor {
 		return NewGossipActor(
 			g.cluster.Config.GossipRequestTimeout,
 			g.cluster.ActorSystem.ID,
@@ -178,11 +181,12 @@ func (g *Gossiper) StartGossiping() error {
 			},
 			g.cluster.Config.GossipFanOut,
 			g.cluster.Config.GossipMaxSend,
+			system,
 		)
 	}), g.GossipActorName)
 
 	if err != nil {
-		plog.Error("Failed to start gossip actor", log.Error(err))
+		g.cluster.Logger().Error("Failed to start gossip actor", slog.Any("error", err))
 		return err
 	}
 
@@ -191,7 +195,7 @@ func (g *Gossiper) StartGossiping() error {
 			g.cluster.ActorSystem.Root.Send(g.pid, topology)
 		}
 	})
-	plog.Info("Started Cluster Gossip")
+	g.cluster.Logger().Info("Started Cluster Gossip")
 	g.throttler = actor.NewThrottle(3, 60*time.Second, g.throttledLog)
 	go g.gossipLoop()
 
@@ -203,40 +207,96 @@ func (g *Gossiper) Shutdown() {
 		return
 	}
 
-	plog.Info("Shutting down gossip")
-	g.cluster.ActorSystem.Root.Stop(g.pid)
-	plog.Info("Shut down gossip")
+	g.cluster.Logger().Info("Shutting down gossip")
+
+	close(g.close)
+
+	err := g.cluster.ActorSystem.Root.StopFuture(g.pid).Wait()
+	if err != nil {
+		g.cluster.Logger().Error("failed to stop gossip actor", slog.Any("error", err))
+	}
+
+	g.cluster.Logger().Info("Shut down gossip")
 }
 
 func (g *Gossiper) gossipLoop() {
-	plog.Info("Starting gossip loop")
+	g.cluster.Logger().Info("Starting gossip loop")
 
 	// create a ticker that will tick each GossipInterval milliseconds
 	// we do not use sleep as sleep puts the goroutine out of the scheduler
-	// P and we do not want our Gs to be scheduled out from the running Ms
+	// P, and we do not want our Gs to be scheduled out from the running Ms
 	ticker := time.NewTicker(g.cluster.Config.GossipInterval)
 breakLoop:
-	for {
+	for !g.cluster.ActorSystem.IsStopped() {
 		select {
 		case <-g.close:
-			plog.Info("Stopping Gossip Loop")
+			g.cluster.Logger().Info("Stopping Gossip Loop")
 			break breakLoop
 		case <-ticker.C:
+
+			g.blockExpiredHeartbeats()
+			g.blockGracefullyLeft()
+
 			g.SetState(HearthbeatKey, &MemberHeartbeat{
 				// todo collect the actor statistics
 				ActorStatistics: &ActorStatistics{},
 			})
 			g.SendState()
-
-			/*
-			 await BlockExpiredHeartbeats();
-
-			 await BlockGracefullyLeft();
-			*/
 		}
 	}
 }
 
+// blockExpiredHeartbeats blocks members that have not sent a heartbeat for a long time
+func (g *Gossiper) blockExpiredHeartbeats() {
+	if g.cluster.Config.GossipInterval == 0 {
+		return
+	}
+	t, err := g.GetState(HearthbeatKey)
+	if err != nil {
+		g.cluster.Logger().Error("Could not get heartbeat state", slog.Any("error", err))
+		return
+	}
+
+	blockList := remote.GetRemote(g.cluster.ActorSystem).BlockList()
+
+	blocked := make([]string, 0)
+
+	for k, v := range t {
+		if k != g.cluster.ActorSystem.ID &&
+			!blockList.IsBlocked(k) &&
+			time.Now().Sub(time.UnixMilli(v.LocalTimestampUnixMilliseconds)) > g.cluster.Config.HeartbeatExpiration {
+			blocked = append(blocked, k)
+		}
+	}
+
+	if len(blocked) > 0 {
+		g.cluster.Logger().Info("Blocking members due to expired heartbeat", slog.String("members", strings.Join(blocked, ",")))
+		blockList.Block(blocked...)
+	}
+}
+
+// blockGracefullyLeft blocking members due to gracefully leaving
+func (g *Gossiper) blockGracefullyLeft() {
+	t, err := g.GetState(GracefullyLeftKey)
+	if err != nil {
+		g.cluster.Logger().Error("Could not get gracefully left members", slog.Any("error", err))
+		return
+	}
+
+	blockList := remote.GetRemote(g.cluster.ActorSystem).BlockList()
+
+	gracefullyLeft := make([]string, 0)
+	for k := range t {
+		if !blockList.IsBlocked(k) && k != g.cluster.ActorSystem.ID {
+			gracefullyLeft = append(gracefullyLeft, k)
+		}
+	}
+	if len(gracefullyLeft) > 0 {
+		g.cluster.Logger().Info("Blocking members due to gracefully leaving", slog.String("members", strings.Join(gracefullyLeft, ",")))
+		blockList.Block(gracefullyLeft...)
+	}
+}
+
 func (g *Gossiper) throttledLog(counter int32) {
-	plog.Debug(fmt.Sprintf("[Gossiper] Gossiper Setting State to %s", g.pid), log.Int("throttled", int(counter)))
+	g.cluster.Logger().Debug(fmt.Sprintf("[Gossiper] Gossiper Setting State to %s", g.pid), slog.Int("throttled", int(counter)))
 }
