@@ -4,21 +4,20 @@ package cluster
 
 import (
 	"fmt"
+	"log/slog"
 	"math/rand"
 	"reflect"
 	"time"
 
 	"github.com/asynkron/gofun/set"
-	"google.golang.org/protobuf/types/known/anypb"
-
-	"github.com/keecon/protoactor-go/actor"
-	"github.com/keecon/protoactor-go/log"
+	"github.com/asynkron/protoactor-go/actor"
 	"google.golang.org/protobuf/proto"
 )
 
 const (
-	TopologyKey   string = "topology"
-	HearthbeatKey string = "heathbeat"
+	TopologyKey       string = "topology"
+	HearthbeatKey     string = "heathbeat"
+	GracefullyLeftKey string = "left"
 )
 
 // create and seed a pseudo random numbers generator
@@ -37,6 +36,7 @@ type Informer struct {
 	gossipFanOut      int
 	gossipMaxSend     int
 	throttler         actor.ShouldThrottle
+	logger            *slog.Logger
 }
 
 // makes sure Informer complies with the Gossip interface
@@ -44,7 +44,7 @@ var _ Gossip = (*Informer)(nil)
 
 // Creates a new Informer value with the given properties and returns
 // back a pointer to its memory location in the heap
-func newInformer(myID string, getBlockedMembers func() set.Set[string], fanOut int, maxSend int) *Informer {
+func newInformer(myID string, getBlockedMembers func() set.Set[string], fanOut int, maxSend int, logger *slog.Logger) *Informer {
 	informer := Informer{
 		myID: myID,
 		state: &GossipState{
@@ -57,6 +57,7 @@ func newInformer(myID string, getBlockedMembers func() set.Set[string], fanOut i
 		getBlockedMembers: getBlockedMembers,
 		gossipFanOut:      fanOut,
 		gossipMaxSend:     maxSend,
+		logger:            logger,
 	}
 	informer.throttler = actor.NewThrottle(3, 60*time.Second, informer.throttledLog)
 	return &informer
@@ -97,7 +98,7 @@ func (inf *Informer) SetState(key string, message proto.Message) {
 	//}
 
 	if _, ok := inf.state.Members[inf.myID]; !ok {
-		plog.Error("State corrupt")
+		inf.logger.Error("State corrupt")
 	}
 
 	inf.checkConsensusKey(key)
@@ -237,12 +238,12 @@ func (inf *Informer) RemoveConsensusCheck(id string) {
 
 // retrieves this informer current state for the given key
 // returns map containing each known member id and their value
-func (inf *Informer) GetState(key string) map[string]*anypb.Any {
-	entries := make(map[string]*anypb.Any)
+func (inf *Informer) GetState(key string) map[string]*GossipKeyValue {
+	entries := make(map[string]*GossipKeyValue)
 
 	for memberID, memberState := range inf.state.Members {
 		if value, ok := memberState.Values[key]; ok {
-			entries[memberID] = value.Value
+			entries[memberID] = value
 		}
 	}
 
@@ -289,5 +290,5 @@ func (inf *Informer) commitPendingOffsets(offsets map[string]int64) {
 }
 
 func (inf *Informer) throttledLog(counter int32) {
-	plog.Debug("[Gossip] Setting State", log.Int("throttled", int(counter)))
+	inf.logger.Debug("[Gossip] Setting State", slog.Int("throttled", int(counter)))
 }
