@@ -2,6 +2,7 @@ package cluster
 
 import (
 	"fmt"
+	"log/slog"
 	"sync"
 	"testing"
 
@@ -20,7 +21,7 @@ func TestInformer_SetState(t *testing.T) {
 		ActorStatistics: &ActorStatistics{},
 	}
 
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 }
 
@@ -35,7 +36,7 @@ func TestInformer_GetState(t *testing.T) {
 		ActorStatistics: &ActorStatistics{},
 	}
 
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 
 	m := i.GetState("heartbeat")
@@ -47,7 +48,7 @@ func TestInformer_GetState(t *testing.T) {
 	}
 
 	var s2 MemberHeartbeat
-	err := x.UnmarshalTo(&s2)
+	err := x.Value.UnmarshalTo(&s2)
 	if err != nil {
 		t.Error("unmarshal state error")
 	}
@@ -65,7 +66,7 @@ func TestInformer_ReceiveState(t *testing.T) {
 	}
 	dummyValue, _ := anypb.New(s)
 
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 
 	remoteState := &GossipState{
@@ -95,8 +96,7 @@ func TestInformer_ReceiveState(t *testing.T) {
 
 	var s1 MemberHeartbeat
 
-	err := m1.UnmarshalTo(&s1)
-
+	err := m1.Value.UnmarshalTo(&s1)
 	if err != nil {
 		t.Error("unmarshal member1 state error")
 	}
@@ -110,12 +110,11 @@ func TestInformer_ReceiveState(t *testing.T) {
 
 	var s2 MemberHeartbeat
 
-	err = m2.UnmarshalTo(&s2)
+	err = m2.Value.UnmarshalTo(&s2)
 
 	if err != nil {
 		t.Error("unmarshal member2 state error")
 	}
-
 }
 
 func TestInformer_SendState(t *testing.T) {
@@ -136,7 +135,7 @@ func TestInformer_SendState(t *testing.T) {
 		ActorStatistics: &ActorStatistics{},
 	}
 
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 	// the cluster sees two nodes. itself and member2
 	i.UpdateClusterTopology(&ClusterTopology{
@@ -169,7 +168,7 @@ func TestInformer_UpdateClusterTopology(t *testing.T) {
 	s := &MemberHeartbeat{
 		ActorStatistics: &ActorStatistics{},
 	}
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 	// the cluster sees two nodes. itself and member2
 	i.UpdateClusterTopology(&ClusterTopology{
@@ -201,7 +200,7 @@ func TestInformer_GetMemberStateDelta(t *testing.T) {
 		ActorStatistics: &ActorStatistics{},
 	}
 
-	i := newInformer("member1", a, 3, 3)
+	i := newInformer("member1", a, 3, 3, slog.Default())
 	i.SetState("heartbeat", s)
 
 	m := i.GetMemberStateDelta("member1")
