@@ -7,6 +7,7 @@ package cluster
 import (
 	"hash"
 	"hash/fnv"
+	"strings"
 	"sync"
 )
 
@@ -15,9 +16,10 @@ type memberData struct {
 	hashBytes []byte
 }
 type Rendezvous struct {
-	mutex   sync.RWMutex
-	hasher  hash.Hash32
-	members []*memberData
+	mutex      sync.RWMutex
+	hasher     hash.Hash32
+	hasherLock sync.Mutex
+	members    []*memberData
 }
 
 func NewRendezvous() *Rendezvous {
@@ -65,38 +67,12 @@ func (r *Rendezvous) GetByClusterIdentity(ci *ClusterIdentity) string {
 }
 
 func (r *Rendezvous) GetByIdentity(identity string) string {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
+	parts := strings.SplitN(identity, "/", 2)
 
-	m := r.members
-	l := len(m)
-
-	if l == 0 {
-		return ""
-	}
-
-	if l == 1 {
-		return m[0].member.Address()
-	}
-
-	keyBytes := []byte(identity)
-
-	var maxScore uint32
-	var maxMember *memberData
-	var score uint32
-
-	for _, node := range m {
-		score = r.hash(node.hashBytes, keyBytes)
-		if score > maxScore {
-			maxScore = score
-			maxMember = node
-		}
-	}
-
-	if maxMember == nil {
-		return ""
-	}
-	return maxMember.member.Address()
+	return r.GetByClusterIdentity(&ClusterIdentity{
+		Kind:     parts[0],
+		Identity: parts[1],
+	})
 }
 
 func (r *Rendezvous) memberDataByKind(kind string) []*memberData {
@@ -126,6 +102,9 @@ func (r *Rendezvous) UpdateMembers(members Members) {
 }
 
 func (r *Rendezvous) hash(node, key []byte) uint32 {
+	r.hasherLock.Lock()
+	defer r.hasherLock.Unlock()
+
 	r.hasher.Reset()
 	r.hasher.Write(key)
 	r.hasher.Write(node)
