@@ -2,8 +2,9 @@ package goring
 
 import (
 	"fmt"
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestPushPop(t *testing.T) {
