@@ -4,15 +4,11 @@ package metrics
 
 import (
 	"fmt"
+	"log/slog"
 	"sync"
 
-	"github.com/keecon/protoactor-go/log"
-	"go.opentelemetry.io/otel/metric/global"
-	"go.opentelemetry.io/otel/metric/instrument"
-	"go.opentelemetry.io/otel/metric/instrument/asyncint64"
-	"go.opentelemetry.io/otel/metric/instrument/syncfloat64"
-	"go.opentelemetry.io/otel/metric/instrument/syncint64"
-	"go.opentelemetry.io/otel/metric/unit"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/metric"
 )
 
 const LibName string = "protoactor"
@@ -25,132 +21,133 @@ type ActorMetrics struct {
 	ID string
 
 	// Actors
-	ActorFailureCount            asyncint64.Counter
-	ActorMailboxLength           asyncint64.Gauge
-	ActorMessageReceiveHistogram syncfloat64.Histogram
-	ActorRestartedCount          asyncint64.Counter
-	ActorSpawnCount              asyncint64.Counter
-	ActorStoppedCount            asyncint64.Counter
+	ActorFailureCount            metric.Int64Counter
+	ActorMailboxLength           metric.Int64ObservableGauge
+	ActorMessageReceiveHistogram metric.Float64Histogram
+	ActorRestartedCount          metric.Int64Counter
+	ActorSpawnCount              metric.Int64Counter
+	ActorStoppedCount            metric.Int64Counter
 
 	// Deadletters
-	DeadLetterCount       asyncint64.Counter
-	FuturesCompletedCount asyncint64.Counter
+	DeadLetterCount metric.Int64Counter
 
 	// Futures
-	FuturesStartedCount  asyncint64.Counter
-	FuturesTimedOutCount asyncint64.Counter
+	FuturesStartedCount   metric.Int64Counter
+	FuturesCompletedCount metric.Int64Counter
+	FuturesTimedOutCount  metric.Int64Counter
 
 	// Threadpool
-	ThreadPoolLatency syncint64.Histogram
+	ThreadPoolLatency metric.Int64Histogram
 }
 
 // NewActorMetrics creates a new ActorMetrics value and returns a pointer to it
-func NewActorMetrics() *ActorMetrics {
-	instruments := newInstruments()
+func NewActorMetrics(logger *slog.Logger) *ActorMetrics {
+	instruments := newInstruments(logger)
 	return instruments
 }
 
 // newInstruments will create instruments using a meter from
 // the given provider p
-func newInstruments() *ActorMetrics {
-	meter := global.Meter(LibName)
+func newInstruments(logger *slog.Logger) *ActorMetrics {
+	meter := otel.Meter(LibName)
 	instruments := ActorMetrics{mu: &sync.Mutex{}}
 
 	var err error
-	if instruments.ActorFailureCount, err = meter.AsyncInt64().Counter(
+
+	if instruments.ActorFailureCount, err = meter.Int64Counter(
 		"protoactor_actor_failure_count",
-		instrument.WithDescription("Number of actor failures"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of actor failures"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ActorFailureCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.ActorMessageReceiveHistogram, err = meter.SyncFloat64().Histogram(
+	if instruments.ActorMessageReceiveHistogram, err = meter.Float64Histogram(
 		"protoactor_actor_message_receive_duration_seconds",
-		instrument.WithDescription("Actor's messages received duration in seconds"),
+		metric.WithDescription("Actor's messages received duration in seconds"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ActorMessageReceiveHistogram instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.ActorRestartedCount, err = meter.AsyncInt64().Counter(
+	if instruments.ActorRestartedCount, err = meter.Int64Counter(
 		"protoactor_actor_restarted_count",
-		instrument.WithDescription("Number of actors restarts"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of actors restarts"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ActorRestartedCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.ActorStoppedCount, err = meter.AsyncInt64().Counter(
+	if instruments.ActorStoppedCount, err = meter.Int64Counter(
 		"protoactor_actor_stopped_count",
-		instrument.WithDescription("Number of actors stopped"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of actors stopped"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ActorStoppedCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.ActorSpawnCount, err = meter.AsyncInt64().Counter(
+	if instruments.ActorSpawnCount, err = meter.Int64Counter(
 		"protoactor_actor_spawn_count",
-		instrument.WithDescription("Number of actors spawn"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of actors spawn"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ActorSpawnCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.DeadLetterCount, err = meter.AsyncInt64().Counter(
+	if instruments.DeadLetterCount, err = meter.Int64Counter(
 		"protoactor_deadletter_count",
-		instrument.WithDescription("Number of deadletters"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of deadletters"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create DeadLetterCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.FuturesCompletedCount, err = meter.AsyncInt64().Counter(
+	if instruments.FuturesCompletedCount, err = meter.Int64Counter(
 		"protoactor_futures_completed_count",
-		instrument.WithDescription("Number of futures completed"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of futures completed"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create FuturesCompletedCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.FuturesStartedCount, err = meter.AsyncInt64().Counter(
+	if instruments.FuturesStartedCount, err = meter.Int64Counter(
 		"protoactor_futures_started_count",
-		instrument.WithDescription("Number of futures started"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of futures started"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create FuturesStartedCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.FuturesTimedOutCount, err = meter.AsyncInt64().Counter(
+	if instruments.FuturesTimedOutCount, err = meter.Int64Counter(
 		"protoactor_futures_timed_out_count",
-		instrument.WithDescription("Number of futures timed out"),
-		instrument.WithUnit(unit.Dimensionless),
+		metric.WithDescription("Number of futures timed out"),
+		metric.WithUnit("1"),
 	); err != nil {
 		err = fmt.Errorf("failed to create FuturesTimedOutCount instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
-	if instruments.ThreadPoolLatency, err = meter.SyncInt64().Histogram(
+	if instruments.ThreadPoolLatency, err = meter.Int64Histogram(
 		"protoactor_thread_pool_latency_duration_seconds",
-		instrument.WithDescription("History of latency in second"),
-		instrument.WithUnit(unit.Milliseconds),
+		metric.WithDescription("History of latency in second"),
+		metric.WithUnit("ms"),
 	); err != nil {
 		err = fmt.Errorf("failed to create ThreadPoolLatency instrument, %w", err)
-		plog.Error(err.Error(), log.Error(err))
+		logger.Error(err.Error(), slog.Any("error", err))
 	}
 
 	return &instruments
 }
 
 // SetActorMailboxLengthGauge makes sure access to ActorMailboxLength is sequenced
-func (am *ActorMetrics) SetActorMailboxLengthGauge(gauge asyncint64.Gauge) {
+func (am *ActorMetrics) SetActorMailboxLengthGauge(gauge metric.Int64ObservableGauge) {
 	// lock our mutex
 	am.mu.Lock()
 	defer am.mu.Unlock()
