@@ -6,9 +6,8 @@ import (
 
 	"github.com/keecon/protoactor-go/actor"
 
-	"github.com/keecon/protoactor-go/internal/queue/goring"
-	"github.com/keecon/protoactor-go/internal/queue/mpsc"
-	"github.com/keecon/protoactor-go/log"
+	"github.com/asynkron/protoactor-go/internal/queue/goring"
+	"github.com/asynkron/protoactor-go/internal/queue/mpsc"
 )
 
 const (
@@ -80,7 +79,6 @@ func (m *endpointWriterMailbox) run() {
 	var msg interface{}
 	defer func() {
 		if r := recover(); r != nil {
-			plog.Info("[ACTOR] Recovering", log.Object("actor", m.invoker), log.Object("reason", r), log.Stack())
 			m.invoker.EscalateFailure(r, msg)
 		}
 	}()
